requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.5
django-taggit==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.4
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
